fix file button popover sourceRect
