+++ Wire-iOS/Sources/UserInterface/Conversation/InputBar/ConversationInputBarViewController+Files.h
@@ -24,7 +24,7 @@ FOUNDATION_EXPORT NSTimeInterval const ConversationUploadMaxVideoDuration;
 
 @interface ConversationInputBarViewController (Files) <UIDocumentMenuDelegate, UIDocumentPickerDelegate, UIImagePickerControllerDelegate, UINavigationControllerDelegate>
 
-- (void)docUploadPressed:(id)sender;
+- (void)docUploadPressed:(IconButton *)sender;
 - (void)uploadFileAtURL:(NSURL *)URL;
 - (void)executeWithVideoPermissions:(dispatch_block_t)toExecute;
 @end
